Implement User–Farm OneToMany mapping and update services/controllers

+++ src/main/java/com/tarsem/khetBuddy_backend/dto/MlRequest.java
@@ -23,7 +23,7 @@ public class MlRequest {
     private double soilPh;
 
     @JsonProperty("soil_moisture")
-    private double soilMoisture;
+    private float soilMoisture;
 
     @JsonProperty("irrigation_type")
     private String irrigationType;
